HGI-9865: properly migrate to hotglue-singer-sdk
